Continue....Add url path, add function to music/views file, create music/urls.py file
Add the music app file so when someone goes to my website it will go to the main urls then go to the music url if music is typed at the end of the url address. The function index(request) with a request parameter is making a http response which is an html content..'hello world' also 'from django.http import HttpResponse' was added, but before that I made a music\urls.py file or a urls.py file for the music app. Then I add the imports and then the urlpatterns.
So when someone comes to my site and types www.music.com they will first go to the 'website' urls to look for the urlpatterns then next it will see the music path and go to the music urls look at the pattern then go to views file for the music app look at the index function and execute that function when is send an http response